+++ projects/angular-split/src/public_api.ts
@@ -4,6 +4,9 @@
 
 export { AngularSplitModule } from './lib/module'
 export { SplitComponent } from './lib/component/split.component'
+export { SplitGutterDirective, SplitGutterTemplateContext } from './lib/gutter/split-gutter.directive'
+export { SplitGutterDragHandleDirective } from './lib/gutter/split-gutter-drag-handle.directive'
+export { SplitGutterExcludeFromDragDirective } from './lib/gutter/split-gutter-exclude-from-drag.directive'
 export { SplitAreaDirective } from './lib/directive/split-area.directive'
 export { ANGULAR_SPLIT_DEFAULT_OPTIONS } from './lib/angular-split-config.token'
 export * from './lib/interface'