+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+<a name="17.1.1"></a>
+
+# 17.1.1 (2023-11-29)
+
+- **Fix** aria-valuenow incorrectly set to wildcard [#368](https://github.com/angular-split/angular-split/pull/368) thanks to [Harpush](https://github.com/Harpush)!
+
+<a name="17.1.0"></a>
+
+# 17.1.0 (2023-11-27)
+
+- **Feat** support custom gutter template [#364](https://github.com/angular-split/angular-split/pull/364) thanks to [Harpush](https://github.com/Harpush)!
+
+<a name="17.0.0"></a>
+
+# 17.0.0 (2023-11-15)
+
+- **Docs** add gutterAriaLabel to documentation [#364](https://github.com/angular-split/angular-split/pull/364) thanks to [jakubmank2](https://github.com/jakubmank2)!
+- **Chore** add support for Angular 17 [#365](https://github.com/angular-split/angular-split/pull/365)
+
+<a name="16.2.1"></a>
+
+# 16.2.1 (2023-09-05)
+
+- **Fix** change default area size from null to `*` [#358](https://github.com/angular-split/angular-split/pull/358) thanks to [Harpush](https://github.com/Harpush)!
+- **Fix** use correct peerDependencies versions [#355](https://github.com/angular-split/angular-split/pull/355) thanks to [beeman](https://github.com/beeman)!
+- **Chore** remove lodash dependency [#356](https://github.com/angular-split/angular-split/pull/356) thanks to [beeman](https://github.com/beeman)!
+- **Chore** update angular.json to modern standards [#357](https://github.com/angular-split/angular-split/pull/357) thanks to [beeman](https://github.com/beeman)!
+
 <a name="16.2.0"></a>
 
 # 16.2.0 (2023-09-03)